// src/main/webapp/app/entities/ERPLTEDB/client/client.model.ts
import dayjs from 'dayjs/esm';
import { ISociete } from 'app/entities/ERPLTEDB/societe/societe.model';
import { IActivite } from 'app/entities/ERPLTEDB/activite/activite.model';
import { ISecteur } from 'app/entities/ERPLTEDB/secteur/secteur.model';
import { IBanquecl } from 'app/entities/ERPLTEDB/banquecl/banquecl.model';
import { IPoste } from 'app/entities/ERPLTEDB/poste/poste.model';
import { IModereg } from 'app/entities/ERPLTEDB/modereg/modereg.model';
import { ICategorietarif } from 'app/entities/ERPLTEDB/categorietarif/categorietarif.model';
import { IGouvernorat } from 'app/entities/ERPLTEDB/gouvernorat/gouvernorat.model';
import { IRegion } from 'app/entities/ERPLTEDB/region/region.model';
import { IRepresentant } from 'app/entities/ERPLTEDB/representant/representant.model';
import { typelocal } from 'app/entities/enumerations/typelocal.model';
import { regime } from 'app/entities/enumerations/regime.model';
import { typeclient } from 'app/entities/enumerations/typeclient.model';
import { typepatante } from 'app/entities/enumerations/typepatante.model';
import { note } from 'app/entities/enumerations/note.model';

export interface IClient {
  id: string;
  code?: string | null;
  name?: string | null;
  adresse?: string | null;
  typelocal?: typelocal | null;
  matriculef?: string | null;
  tel?: string | null;
  fax?: string | null;
  regcom?: string | null;
  email?: string | null;
  rib?: string | null;
  personnecont?: string | null;
  photo?: string | null;
  photoContentType?: string | null;
  srisque?: number | null;
  scredit?: number | null;
  blocage?: boolean | null;
  timbre?: boolean | null;
  regime?: regime | null;
  typeclient?: typeclient | null;
  typepatante?: typepatante | null;
  note?: note | null;
  numportable?: string | null;
  delaitreg?: number | null;
  numdecision?: string | null;
  datedebut?: dayjs.Dayjs | null;
  datefin?: dayjs.Dayjs | null;
  datecreation?: dayjs.Dayjs | null;
  datevalidation?: dayjs.Dayjs | null;
  societe?: Pick<ISociete, 'id' | 'code'> | null;
  activite?: Pick<IActivite, 'id' | 'name'> | null;
  secteur?: Pick<ISecteur, 'id' | 'name'> | null;
  banquecl?: Pick<IBanquecl, 'id' | 'name'> | null;
  poste?: Pick<IPoste, 'id' | 'code'> | null;
  modereg?: Pick<IModereg, 'id' | 'name'> | null;
  categorietarif?: Pick<ICategorietarif, 'id' | 'name'> | null;
  gouvernorat?: Pick<IGouvernorat, 'id' | 'name'> | null;
  region?: Pick<IRegion, 'id' | 'name'> | null;
  representant?: Pick<IRepresentant, 'id' | 'name'> | null;
}

export type NewClient = Omit<IClient, 'id'> & { id: null };